requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
